pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Lucas Ferreira <lucasferreiraek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic = {extras = ["email"], version = "^2.5.2"}
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.1.7"
httpx = "^0.25.2"
isort = "^5.12.0"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations', '/home/$USER/.cache/']

[tool.isort]
profile = 'black'
line-length = 79
extend_skip = ['migrations', '/home/$USER/.cache/']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend_exclude = '(migrations/, .*/, /home/$USER/.cache/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn fast_zero.app:app --reload'
pre-test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post-test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
